# docs/swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample API to manage users"
  version: "1.0"
  title: "User API"
  contact:
    name: "API Support"
    url: ""
    email: ""
basePath: "/api/v1"
paths:
  /user:
    post:
      summary: "Create a user"
      description: "Create a new user with name, surname, patronymic, age, gender, nationality"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "user"
          description: "Create User"
          required: true
          schema:
            type: "object"
            properties:
              name:
                type: "string"
              surname:
                type: "string"
              patronymic:
                type: "string"
              age:
                type: "integer"
                format: "int32"
              gender:
                type: "string"
              nationality:
                type: "string"
      responses:
        200:
          description: "User created successfully"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
                format: "int64"
              name:
                type: "string"
              surname:
                type: "string"
              patronymic:
                type: "string"
              age:
                type: "integer"
                format: "int32"
              gender:
                type: "string"
              nationality:
                type: "string"
        400:
          description: "Bad request"
          schema:
            type: "object"
            properties:
              message:
                type: "string"

  /users:
    get:
      summary: "Get all users"
      description: "Get a list of users"
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful response"
          schema:
            type: "array"
            items:
              type: "object"
              properties:
                id:
                  type: "integer"
                  format: "int64"
                name:
                  type: "string"
                surname:
                  type: "string"
                patronymic:
                  type: "string"
                age:
                  type: "integer"
                  format: "int32"
                gender:
                  type: "string"
                nationality:
                  type: "string"

  /users/{id}:
    get:
      summary: "Get user by ID"
      description: "Get a user by ID"
      parameters:
        - in: "path"
          name: "id"
          description: "User ID"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "User retrieved successfully"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
                format: "int64"
              name:
                type: "string"
              surname:
                type: "string"
              patronymic:
                type: "string"
              age:
                type: "integer"
                format: "int32"
              gender:
                type: "string"
              nationality:
                type: "string"
        404:
          description: "User not found"
          schema:
            type: "object"
            properties:
              message:
                type: "string"

    put:
      summary: "Update a user"
      description: "Update user details"
      parameters:
        - in: "path"
          name: "id"
          description: "User ID"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "user"
          description: "Update User"
          required: true
          schema:
            type: "object"
            properties:
              name:
                type: "string"
              surname:
                type: "string"
              patronymic:
                type: "string"
              age:
                type: "integer"
                format: "int32"
              gender:
                type: "string"
              nationality:
                type: "string"
      responses:
        200:
          description: "User updated successfully"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
                format: "int64"
              name:
                type: "string"
              surname:
                type: "string"
              patronymic:
                type: "string"
              age:
                type: "integer"
                format: "int32"
              gender:
                type: "string"
              nationality:
                type: "string"
        400:
          description: "Bad request"
          schema:
            type: "object"
            properties:
              message:
                type: "string"

    delete:
      summary: "Delete a user"
      description: "Delete user by ID"
      parameters:
        - in: "path"
          name: "id"
          description: "User ID"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "User deleted successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
        404:
          description: "User not found"
          schema:
            type: "object"
            properties:
              message:
                type: "string"